split stories up into multiple fies
